Add suite file for web and make tck/api versions clear
An extra suite.xml file was added to the starter module to test the
web-profile. The guide was updated to mention this.

Since the API version of Concurrency (e.g. 3.0.1) and the TCK version
(e.g. 3.0.2) don't necessarily have to be the same, the properties for
these were renamed to make it more clear which is which.

Signed-off-by: Arjan Tijms <arjan.tijms@gmail.com>